refactor: split agent create/edit forms, fix access tab, restructure email tabs
- Split AgentEditForm into separate create (single-page, no tabs) and edit
  (tabbed) forms with shared field components in agent-form-fields.tsx
- Fix Access tab "Failed to load access list" — workspace API calls were
  missing workspace_id query param, causing 400 from withWorkspaceMember
- Restructure email page: Inbox (all inbound, green dot for whitelisted),
  Sent, Trusted (whitelisted senders only) — replaces old Rejected tab
- Add Senders tab to agent edit form for whitelist management
- Align workspace settings content width with agent edit (max-w-md)
- Add missing tests for access routes (null ownerId, whitelist removal)
